TESTS: Added -k to keep temp code
Closes #11. Each call to the code runner in the test module
creates a new scratch .py file that can be retained after
tests are run by passing `-k` to `tests.py`. Now the module-
based code running just needs to be rolled out to replace the
`exec`-based runner for the rest of the `test_` methods.

 * Created `tempvars/test/consts.py` and added to `__init__.py`.

 * The `-k` argument was added to the `argparse` code in
`tests.py` in the project root.