DEV: Add shallow copies for names/starts/ends
The objects actually worked with for `names`, `starts`, and
`ends` are now shallow copies of the objects passed in. This
won't make any difference in the major anticipated use case,
where list literals are passed, but in the event that someone
constructs `names` programmatically or uses a variable to hold
a common set of variable names, it now won't end up with
a bunch of other variables tagged into it that matched `starts`
or `ends`.

The shallow copying of `starts` and `ends` probably isn't needed now,
since they aren't `.append()`-ed to, but better to protect them now
in case this should change later on.